remove contextPatcher
